mesurement input を整形
