+++ common/styles/_utility_900up.scss
@@ -3,10 +3,6 @@
 Utility Style : base
 
 ******************************************************************/
-body {
-    @import url('https://fonts.googleapis.com/css?family=Noto+Sans|Noto+Sans+JP&display=swap');
-    font-family: "Noto Sans", "Noto Sans JP";
-}
 
 div.mydialog_header {
     height: 60px;
