+++ common/styles/_utility_base.scss
@@ -3,6 +3,8 @@
 Utility Style : base
 
 ******************************************************************/
+@charset "utf-8";
+
 html, body,p, div,ul,li {
     margin: 0;
     padding: 0;
