+++ health/input/styles/breakpoints/_base.scss
@@ -1,15 +1,3 @@
-@charset "utf-8";
-
-/* **************************
-    fonts
-   ************************** */
-
-@import url("https://fonts.googleapis.com/css?family=Noto+Sans|Noto+Sans+JP&display=swap");
-
-/* **************************
-    モバイル用設定
-   ************************** */
-
 /* **************************
     基本設定
    ************************** */
@@ -91,83 +79,3 @@ div.mydialog {
     
 
 
-
-/* ***************************************
-    Settings for PC
-   ************************************ */
-
-@media screen and (min-width: 500px) {
-  header {
-    width: 100%;
-    height: 60px;
-    line-height: 60px;
-  }
-
-  div.headercontainer {
-    width: 300px;
-    height: 40px;
-    font-size: 40px;
-  }
-
-  #homeicon {
-    font-size: 30px;
-  }
-}
-
-/* **************************
-    mycontainer
-   ************************** */
-
-@media screen and (min-width: 500px) {
-  div.mycontainer {
-    max-width: 990px;
-    margin: 0 auto;
-    height: 100vh;
-    padding-top: 20px;
-    background-color: rgba(240, 240, 240, 0.5);
-    /* text-align: center; */
-  }
-
-  .form-group {
-    width: 600px;
-    margin: 0 auto;
-  }
-}
-/* **************************
-    card
-   ************************** */
-
-@media screen and (min-width: 500px) {
-  div.mycard {
-    width: 700px;
-    height: auto;
-    margin: 0 auto;
-    margin-top: 40px;
-  }
-
-  div.mycardheader {
-    width: 100%;
-    height: 100px;
-    margin-bottom: 30px;
-  }
-
-  div.mycardheader p {
-    height: 50px;
-    font-size: 2em;
-  }
-
-  form.select {
-    display: block;
-  }
-
-  #btn_submit {
-    width: 100%;
-    height: 50px;
-    margin: 20px 0;
-  }
-  #btn_cancel {
-    width: 100%;
-    height: 40px;
-    /* margin-bottom: 20px; */
-  }
-}
