+++ health/input/styles/style.css
@@ -5,8 +5,6 @@ Utility Style : base
 
 ******************************************************************/
 @import url("https://fonts.googleapis.com/css?family=Noto+Sans|Noto+Sans+JP&display=swap");
-@import url("https://fonts.googleapis.com/css?family=Noto+Sans|Noto+Sans+JP&display=swap");
-@import url("https://fonts.googleapis.com/css?family=Noto+Sans|Noto+Sans+JP&display=swap");
 html, body, p, div, ul, li {
   margin: 0;
   padding: 0;
@@ -60,12 +58,6 @@ div.mydialog_header p {
   background-color: lightgray;
 }
 
-/* **************************
-    fonts
-   ************************** */
-/* **************************
-    モバイル用設定
-   ************************** */
 /* **************************
     基本設定
    ************************** */
@@ -142,77 +134,6 @@ div.mydialog {
   color: blueviolet;
 }
 
-/* ***************************************
-    Settings for PC
-   ************************************ */
-@media screen and (min-width: 500px) {
-  header {
-    width: 100%;
-    height: 60px;
-    line-height: 60px;
-  }
-  div.headercontainer {
-    width: 300px;
-    height: 40px;
-    font-size: 40px;
-  }
-  #homeicon {
-    font-size: 30px;
-  }
-}
-
-/* **************************
-    mycontainer
-   ************************** */
-@media screen and (min-width: 500px) {
-  div.mycontainer {
-    max-width: 990px;
-    margin: 0 auto;
-    height: 100vh;
-    padding-top: 20px;
-    background-color: rgba(240, 240, 240, 0.5);
-    /* text-align: center; */
-  }
-  .form-group {
-    width: 600px;
-    margin: 0 auto;
-  }
-}
-
-/* **************************
-    card
-   ************************** */
-@media screen and (min-width: 500px) {
-  div.mycard, div.mydialog {
-    width: 700px;
-    height: auto;
-    margin: 0 auto;
-    margin-top: 40px;
-  }
-  div.mycardheader {
-    width: 100%;
-    height: 100px;
-    margin-bottom: 30px;
-  }
-  div.mycardheader p {
-    height: 50px;
-    font-size: 2em;
-  }
-  form.select {
-    display: block;
-  }
-  #btn_submit {
-    width: 100%;
-    height: 50px;
-    margin: 20px 0;
-  }
-  #btn_cancel {
-    width: 100%;
-    height: 40px;
-    /* margin-bottom: 20px; */
-  }
-}
-
 @media (min-width: 500px) {
   /* **************************
     mycontainer
@@ -243,9 +164,6 @@ div.mydialog {
 Utility Style : base
 
 ******************************************************************/
-  body {
-    font-family: "Noto Sans", "Noto Sans JP";
-  }
   div.mydialog_header {
     height: 60px;
     margin-bottom: 20px;
